+++ src/App.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+function App() {
+  return (
+    <div className="App">
+      <h1>Hello World</h1>
+    </div>
+  );
+}
+
+export default App;
